+++ src/redux/modules/entities/roles/selectors.js
@@ -7,27 +7,31 @@ import { Map } from 'immutable';
 import { getSelectedEntity } from '../selectors';
 import { selectFormInitialValues } from '../../form/selectors';
 
-export const getRoles = state => {
-  return state.getIn(['Entities', 'roles'], new Map([]));
-};
+export const getRoles = state => state.getIn(['Entities', 'roles'], new Map([]));
 
-export const convertRoles = createSelector([getRoles], roles => roles.get('data').toJS());
+export const convertRoles = createSelector(
+  [getRoles],
+  roles => roles.get('data').toJS()
+);
 
-export const selectTenantRoles = createSelector([getRoles], roles => {
-  const rolesArray = roles
-    .get('data')
-    .toJS()
-    .filter(role => role.active)
-    .map(role => ({
-      value: role.id,
-      label: role.name
-    }));
-  if (rolesArray.length > 0) {
-    rolesArray.splice(0, 0, rolesArray[2]);
-    rolesArray.splice(3, 1);
+export const selectTenantRoles = createSelector(
+  [getRoles],
+  roles => {
+    const rolesArray = roles
+      .get('data')
+      .toJS()
+      .filter(role => role.active)
+      .map(role => ({
+        value: role.id,
+        label: role.name
+      }));
+    if (rolesArray.length > 0) {
+      rolesArray.splice(0, 0, rolesArray[2]);
+      rolesArray.splice(3, 1);
+    }
+    return rolesArray;
   }
-  return rolesArray;
-});
+);
 export const selectFirstTenantRoleValue = state => {
   const roles = selectTenantRoles(state);
   if (roles && roles.length > 0) {
@@ -37,14 +41,16 @@ export const selectFirstTenantRoleValue = state => {
 
 export const getPlatformRoles = state => state.getIn(['Entities', 'platformRoles', 'data']);
 
-export const selectPlatformRoles = createSelector([getPlatformRoles], roles =>
-  roles
-    .toJS()
-    .map(role => ({
-      value: role.id,
-      label: role.name
-    }))
-    .reverse()
+export const selectPlatformRoles = createSelector(
+  [getPlatformRoles],
+  roles =>
+    roles
+      .toJS()
+      .map(role => ({
+        value: role.id,
+        label: role.name
+      }))
+      .reverse()
 );
 
 export const selectFirstPlatformRoleValue = state => {